Move previous/next reference code from editor.js to panel-reference-font.js
